DROOLS-5128: Update com.google.elemental2 and com.google.jsinterop versions (#3256)
For more details see https://issues.redhat.com/browse/DROOLS-5128

+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/test/java/org/kie/workbench/common/dmn/client/editors/types/common/ScrollHelperTest.java
@@ -91,7 +91,7 @@ public class ScrollHelperTest {
     public void testAnimatedScrollToBottomWithOneArgument() {
 
         final Element element = mock(Element.class);
-        final double scrollHeight = 123;
+        final int scrollHeight = 123;
 
         element.scrollHeight = scrollHeight;
         doNothing().when(scrollHelper).animatedScrollToBottom(any(), anyDouble());
